+++ apps/strapi-cms/CHANGELOG.md
@@ -1,5 +1,12 @@
 # strapi-cms
 
+## 0.2.0
+
+### Minor Changes
+
+- 1e4b96e: Set logo as required in product schema
+- 68ac9bc: Add Product schema and productsShowcase component to homepage in Strapi CMS
+
 ## 0.1.0
 
 ### Minor Changes